Demonstrate Math.toIntExact and Long.intValue
Demonstrates use of Math.toIntExact(Long) to convert Long to int and contrasts this with use of Long.intValue() to do the same. This code was originally cited in the blog post "Exact Conversion of Long to int in Java" (http://marxsoftware.blogspot.com/exact-conversion-of-long-to-int-in-java.html) on the "Inspired by Actual Events" blog (http://marxsoftware.blogspot.com/).