Added three more demonstrations
Added demonstrations of NPEs thrown. First added demonstration uses series of method invocations in which any one (or more) of those methods could be the culprit returning null. Second added demonstration throws NPE when one of the null arguments to a constructor is automatically dereferenced to fit the primitive the method expects. The third added demonstration is the same as the second, but is with a method instead of a constructor.